requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bibtexparser==1.4.1
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
docutils==0.20.1
email_validator==2.1.1
executing==2.0.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.14.0
flatbuffers==24.3.25
free-proxy==1.1.1
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
imagesize==1.4.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kubernetes==29.0.0
langchain==0.1.16
langchain-chroma==0.1.0
langchain-community==0.0.34
langchain-core==0.1.46
langchain-google-genai==1.0.3
langchain-google-vertexai==1.0.2
langchain-openai==0.1.4
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.51
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.24.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
parso==0.8.4
pillow==10.3.0
pinecone-client==3.2.2
platformdirs==4.2.1
posthog==3.5.0
praw==7.7.1
prawcore==2.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psaw==0.1.0
psutil==5.9.8
pulsar-client==3.5.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
# pywin32==3.0.6
PyYAML==6.0.1
pyzmq==26.0.2
reddit-post-scraping-tool==2.0.0.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rich-argparse==1.4.0
rsa==4.9
safetensors==0.4.3
scholarly==1.7.11
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.20.0
sentence-transformers==2.7.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-protobuf==4.25.0.20240417
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
ujson==5.9.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
